Fixing Content::Html outputting as an array bug.
Iconv returns an array, and because of changes to Array#to_s in Ruby 1.9
this results in outputting the string as an array: ["<p>Hello</p>"]

The solution is to simply use Array#first to pull the string out of the
array since there will only ever be one element in the array. This
should work for Ruby 1.8 and 1.9.