support external content references
a <content> element in an Atom entry may have a 'src' attribute that
points to an external location for that content, rather than placing
the content inline (as a child of the <content> element). support
capturing these flavors of content with access to the src attribute.
this is particularly useful for google docs api integration (which is
atom based).