Add different UTF-8 checking for ruby 1.9+.
Instead of iconv (which shows a deprecation warning in 1.9.3), use String.force_encoding(UTF-8).ascii_only? to check for non-utf8 characters and throw an exception instead of passing on to libxml.

Leave the old code there and use it if the ruby version matches 1.8 since 1.8 doesn't have String.force_encoding.